+++ RELEASE_NOTES.md
@@ -7,8 +7,9 @@
 - Changed mouse button callbacks to contain default functions.
 - Added mouse moved, button down, and button up callbacks; changed all mouse calls to default empty function bodies.
 - Optimized timer, box, and text widgets, removed functions that previously needed to be overridden.
-- Added SingleCallback, PointCallback, and ButtonCallback for closures.
+- Added SingleCallback and PointCallback for closures.
 - Added callbacks(&mut self) call to all widgets to allow for callback/closure storage.
+- Added callbacks for mouse enter, exit, scroll, and move.
 
 ## 0.1.12